Reorganize code to prepare for multilingual site
This commit moves the content of the Jekyll site into a separate folder for
the english content (directory "en"). A new index.html file is added which
simply redirects to the English index.html page. Each Jekyll file now has a
'lang' variable in the frontmatter, which allows the navbar in the header to
be generated based on the currently active language. Finally, the footer links
are now defined by variables defined in the layout frontmatter, so that a
single place exists for this language-specific content. I've added Dutch links
there to show how this would work, and rewrote the footer code to get the
content from these variables.